visitor done

+++ routes/web.php
@@ -1,4 +1,6 @@
 <?php
+use Illuminate\Contracts\Encryption\DecryptException;
+use Illuminate\Support\Facades\Crypt;
 
 use App\Http\Controllers\ArtikelController;
 use App\Http\Controllers\HomeController;
@@ -14,7 +16,7 @@ Route::get('/', function () {
 });
 
 Route::get('/', [HomeController::class, 'index'])->name('welcome');
-Route::get('/produk/{id}', [HomeController::class, 'show']);
+Route::get('/produk/{id}', [HomeController::class, 'show'])->name('produk.show');
 
 //kategori
 Route::get('/kategori', [viewkategController::class, 'index'])->name('index');
@@ -29,10 +31,9 @@ Route::get('/kategori/bros-ring', [viewkategController::class, 'brosRing']);
 Route::get('/detailArtikel/{id}', [HomeController::class, 'show1'])->name('Artikel.detail');
 
 
-Route::get('/dashboard', function () {
-    $title = "Dashboard";
-    return view('dashboard', compact('title'));
-})->middleware(['auth', 'verified'])->name('dashboard');
+Route::get('/dashboard', [DashboardController::class, 'index'])
+    ->middleware(['auth', 'verified'])
+    ->name('dashboard');
 
 // Route::get('/dashboard', [dashboardController::class, 'index'])
 //     ->middleware(['auth', 'verified'])
@@ -44,7 +45,7 @@ Route::middleware('auth')->group(function () {
     Route::delete('/profile', [ProfileController::class, 'destroy'])->name('profile.destroy');
 
     //produk
-    Route::get('/produk', function () {
+    Route::get('admin/produk', function () {
         $search = request('search');
         $posts = Produk::latest();
 
@@ -54,15 +55,15 @@ Route::middleware('auth')->group(function () {
 
         $posts = $posts->get();
 
-        return view('produk', ['posts' => $posts, 'search' => $search]);
+        return view('admin/produk', ['posts' => $posts, 'search' => $search]);
     })->name('produk.index');
 
-    Route::get('/produk', [ProdukController::class, 'index'])->name('produk.index');
-    Route::get('/tambah', [ProdukController::class, 'create'])->name('produk.create');
-    Route::post('/produk', [ProdukController::class, 'store'])->name('produk.store');
-    Route::get('/produk/{id}/edit', [ProdukController::class, 'edit'])->name('produk.edit');
-    Route::put('/produk/{id}', [ProdukController::class, 'update'])->name('produk.update');
-    Route::delete('/produk/{id}', [ProdukController::class, 'destroy'])->name('produk.destroy');
+    Route::get('admin/produk', [ProdukController::class, 'index'])->name('produk.index');
+    Route::get('admin/tambahProduk', [ProdukController::class, 'create'])->name('produk.create');
+    Route::post('admin/produk', [ProdukController::class, 'store'])->name('produk.store');
+    Route::get('admin/produk/{id}/edit', [ProdukController::class, 'edit'])->name('produk.edit');
+    Route::put('admin/produk/{id}', [ProdukController::class, 'update'])->name('produk.update');
+    Route::delete('admin/produk/{id}', [ProdukController::class, 'destroy'])->name('produk.destroy');
 
     // Artikel
     Route::get('/artikel', [ArtikelController::class, 'index'])->name('Artikel.index');
